+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Set CAPVCD `ClusterReady` timeout to 25Min in basic test
+
 ## [1.77.0] - 2024-11-19
 
 ### Changed